# docs/swagger.yaml
basePath: /api/v1
definitions:
  api.ClientListResponse:
    properties:
      client_id:
        type: string
      created_at:
        type: string
      description:
        type: string
      name:
        type: string
      redirect_uri:
        type: string
      scopes:
        type: string
      status:
        type: string
    type: object
  api.ClientRegistrationRequest:
    properties:
      description:
        type: string
      name:
        type: string
      redirect_uri:
        type: string
      scopes:
        type: string
      token_ttl_seconds:
        type: integer
    required:
    - name
    - redirect_uri
    - scopes
    type: object
  api.ClientRegistrationResponse:
    properties:
      client_id:
        type: string
      client_secret:
        type: string
      created_at:
        type: string
      description:
        type: string
      name:
        type: string
      redirect_uri:
        type: string
      scopes:
        type: string
      status:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 企业级KYC认证服务
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: KYC Service API
  version: "1.0"
paths:
...
  /clients:
    get:
      consumes:
      - application/json
      description: List OAuth clients in current organization
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.ClientListResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List OAuth clients
      tags:
      - Client Management
  /clients/{client_id}:
    delete:
      consumes:
      - application/json
      description: Delete specified OAuth client
      parameters:
      - description: Client ID
        in: path
        name: client_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete OAuth client
      tags:
      - Client Management
  /clients/register:
    post:
      consumes:
      - application/json
      description: 注册新的OAuth客户端，获取client_id和client_secret
      parameters:
      - description: Client registration request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.ClientRegistrationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.ClientRegistrationResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 注册OAuth客户端
      tags:
      - Client Management
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"